+++ reports-reloaded/src/main/java/com/novoda/github/reports/data/mock/MockProjectDataLayer.java
@@ -3,7 +3,7 @@ package com.novoda.github.reports.data.mock;
 import com.novoda.github.reports.data.ProjectDataLayer;
 import com.novoda.github.reports.data.model.ProjectRepoStats;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 public class MockProjectDataLayer implements ProjectDataLayer {
@@ -11,12 +11,12 @@ public class MockProjectDataLayer implements ProjectDataLayer {
     public ProjectRepoStats getStats(String project, Date from, Date to) {
         return new ProjectRepoStats(
                 project,
-                BigDecimal.valueOf(12),
-                BigDecimal.valueOf(13),
-                BigDecimal.valueOf(48),
-                BigDecimal.valueOf(96),
-                BigDecimal.valueOf(123),
-                BigDecimal.valueOf(8)
+                BigInteger.valueOf(12),
+                BigInteger.valueOf(13),
+                BigInteger.valueOf(48),
+                BigInteger.valueOf(96),
+                BigInteger.valueOf(123),
+                BigInteger.valueOf(8)
         );
     }
 }
